Disable TLS host check on localhost.
This is a more sophisticated solution to this issue
f7154c366fc3cb325465b313675f1954e20048e5.

It turns out that there is a use case in which there is
no other option but to use the SMC: it's when the LRA does
Bootstrapping and it configures the certificates by requesting
them from itself. The SMC of course has no host/IPs in it so
this check breaks Bootstrapping, which means there is no way to
automatically configure the LRA without some other workaround.

These 2 hosts are considered as localhost: "127.0.0.1", "localhost".
The affected function is CMPclient_setup_HTTP() in which call to
STORE_set1_host_ip is skipped.